requirements.txt
matplotlib==1.3.1
numpy==1.15.0
pyserial==3.4
